pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-genai-bridge"
version = "0.1.5"
description = "Native Google GenAI SDK integration for LangGraph with context caching, tool auto-conversion, and bidirectional message translation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Pierre Samson"},
    {name = "Claude Opus (Anthropic)", email = "noreply@anthropic.com"},
]
keywords = [
    "langgraph", "google-genai", "gemini", "langchain", "context-caching",
    "function-calling", "ai-agents", "llm", "orchestration",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "google-genai>=1.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.3.0"]
langchain-google = ["langchain-google-genai>=2.1.0"]
dev = ["pytest>=7.0", "pytest-mock"]
all = [
    "langgraph>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "pytest>=7.0",
]

[project.urls]
Homepage = "https://github.com/darw007d/langgraph-genai-bridge"
Repository = "https://github.com/darw007d/langgraph-genai-bridge"
Issues = "https://github.com/darw007d/langgraph-genai-bridge/issues"

[tool.setuptools.packages.find]
where = ["src"]
